% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignConst.R
\name{assign.const}
\alias{assign.const}
\title{Assign Constants for 6800.  Used by read6800()}
\usage{
assign.const(x, y)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\description{
Function to create varibles for each of the constants found in the header
}
