% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change6400LA.R
\name{change.6400la}
\alias{change.6400la}
\title{Function to change leaf area of LiCor 6400 text files.  Read data into R using read6400() for best results.}
\usage{
change.6400la(x, leaf.area, rm.extra = T)
}
\arguments{
\item{x}{Data frame with licor 6400 data from text files.}

\item{leaf.area}{Numeric. New leaf area in cm^2.}

\item{rm.extra}{Logical. Remove extra columns that were created.  Defaults to True.}
}
\description{
Function to change leaf area of LiCor 6400 text files.  Read data into R using read6400() for best results.
}
