% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change6800LA.R
\name{change.6800la}
\alias{change.6800la}
\title{Change leaf area for Licor 6800 files.  Read data into R using read6800() for best results.}
\usage{
change.6800la(x, leaf.area, Out.file = FALSE)
}
\arguments{
\item{x}{Data frame with licor 6800 data.}

\item{leaf.area}{New leaf area.}

\item{Out.file}{logical. Output a file instead of a dataframe.}
}
\description{
This function allows you to change leaf area of Licor 6800 csv files.
}
\keyword{6800}
\keyword{Licor}
