% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeAbs.R
\name{change.abs}
\alias{change.abs}
\title{Function to change absorbance in licor files}
\usage{
change.abs(x, red.abs, blue.abs, adark = NULL)
}
\arguments{
\item{x}{Data frame with licor 6400 data from text file.}

\item{red.abs}{Numeric. Red absorbance.}

\item{blue.abs}{Numeric. Blue absorbance.}

\item{adark}{Numeric. Dark Assimilation.  Must be negative.}
}
\description{
Function to change absorbance in licor files
}
