% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_QC.R
\name{licor.qc}
\alias{licor.qc}
\title{QA/QC Licor data}
\usage{
licor.qc(
  x,
  out.dir = NULL,
  gsw,
  gsw.cutoff = 0,
  Ci,
  Ci.cutoff = c(0, 2000),
  Tleaf,
  Tleaf.cutoff = NULL,
  QC = NULL
)
}
\arguments{
\item{x}{Data frame. Input data}

\item{out.dir}{Output directory for QC data and info}

\item{gsw}{Character.  Name of column containing conductance values.}

\item{gsw.cutoff}{Numeric. Cutoff for low conductance.}

\item{Ci}{Character.  Name of column containing Ci values.}

\item{Ci.cutoff}{Numeric list. Cutoff range of values in form (low, high) for nonsensical Ci.  Default is c(0, 2000)}

\item{Tleaf}{Character.  Name of column containing Tleaf values. Optional.}

\item{Tleaf.cutoff}{Numeric. Cutoff for indiviual Tleaf variation from mean.}

\item{QC}{Character. Initial QC column name. Optional}
}
\description{
QA/QC Licor data
}
