% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadHeader.R
\name{read.header}
\alias{read.header}
\title{Function to read LiCor 6400 header}
\usage{
read.header(licor.file, start.line = 30, n.rows = NULL)
}
\arguments{
\item{licor.file}{LiCor file.}

\item{start.line}{Numeric.  How many lines to start of data.  Usually set to more than expected.  The more lines, the slower the code.  Defaults to 30.}

\item{n.rows}{Numberic. How many rows to read.  This will be calculated if start.line is provided.}
}
\description{
Function to read LiCor 6400 header
}
