% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read6400.R
\name{read6400}
\alias{read6400}
\title{Function to read LiCor 6400 text files}
\usage{
read6400(licor.file, start.line = 30, time = T, leaf.area = NULL, file.out = F)
}
\arguments{
\item{licor.file}{Path to LiCor file.}

\item{start.line}{Numeric.  How many lines to start of data.  Usually set to more than expected.  The more lines, the slower the code.  Defaults to 30.}

\item{time}{Logical.  Convert time column from char to POSIXct.  Defaults to True.}

\item{leaf.area}{Numeric.  Leaf area inside chamber.  Only required if it needs to be changed.}

\item{file.out}{Logical.  Output a file with new leaf area instead of dataframe.  Output file will look like a 6400 file with remarks removed.  Default is False.}
}
\description{
Function to read LiCor 6400 text files
}
