% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read6800File.R
\name{read6800}
\alias{read6800}
\title{Read Licor 6800 files}
\usage{
read6800(file.name, constants = TRUE, Out.data = TRUE)
}
\arguments{
\item{file.name}{File name of licor file.  If file not in working directory, include path to file.}

\item{constants}{Logical.  If true, create varibles for each of the constants found in the header.  If false, varibles in header will be deleted.}

\item{Out.data}{Logical. If true, outputs data from file.  If false, outputs constants header only.}
}
\description{
Function reads file into R, creates varibles from header constants, changes special characters to plain text, and outputs a dataframe.
}
\keyword{6800}
\keyword{Licor}
